+++ src/main/java/com/ykotsiuba/profitsoft_2/dto/article/SearchArticleRequestDTO.java
@@ -1,4 +1,4 @@
-package com.ykotsiuba.profitsoft_2.dto;
+package com.ykotsiuba.profitsoft_2.dto.article;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
