+++ src/main/java/com/ykotsiuba/profitsoft_2/dto/article/UploadArticlesResponseDTO.java
@@ -1,4 +1,4 @@
-package com.ykotsiuba.profitsoft_2.dto;
+package com.ykotsiuba.profitsoft_2.dto.article;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,6 +9,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class DeleteArticleResponseDTO {
-    String message;
+public class UploadArticlesResponseDTO {
+    private Integer uploaded;
+    private Integer errors;
 }
