+++ src/main/java/com/ykotsiuba/profitsoft_2/mapper/AuthorMapper.java
@@ -1,6 +1,6 @@
 package com.ykotsiuba.profitsoft_2.mapper;
 
-import com.ykotsiuba.profitsoft_2.dto.AuthorDTO;
+import com.ykotsiuba.profitsoft_2.dto.author.AuthorDTO;
 import com.ykotsiuba.profitsoft_2.entity.Author;
 import org.mapstruct.Mapper;
