+++ src/main/java/com/ykotsiuba/profitsoft_2/service/impl/AuthorServiceImpl.java
@@ -1,8 +1,8 @@
 package com.ykotsiuba.profitsoft_2.service.impl;
 
-import com.ykotsiuba.profitsoft_2.dto.AuthorDTO;
-import com.ykotsiuba.profitsoft_2.dto.DeleteAuthorResponseDTO;
-import com.ykotsiuba.profitsoft_2.dto.SaveAuthorRequestDTO;
+import com.ykotsiuba.profitsoft_2.dto.author.AuthorDTO;
+import com.ykotsiuba.profitsoft_2.dto.author.DeleteAuthorResponseDTO;
+import com.ykotsiuba.profitsoft_2.dto.author.SaveAuthorRequestDTO;
 import com.ykotsiuba.profitsoft_2.entity.Author;
 import com.ykotsiuba.profitsoft_2.mapper.AuthorMapper;
 import com.ykotsiuba.profitsoft_2.repository.AuthorRepository;
