+++ src/test/java/com/ykotsiuba/profitsoft_2/controller/AuthorControllerTest.java
@@ -9,9 +9,9 @@ import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.ykotsiuba.profitsoft_2.Profitsoft2Application;
 import com.ykotsiuba.profitsoft_2.TestProfitsoft2Application;
 import com.ykotsiuba.profitsoft_2.dto.APIException;
-import com.ykotsiuba.profitsoft_2.dto.AuthorDTO;
-import com.ykotsiuba.profitsoft_2.dto.DeleteAuthorResponseDTO;
-import com.ykotsiuba.profitsoft_2.dto.SaveAuthorRequestDTO;
+import com.ykotsiuba.profitsoft_2.dto.author.AuthorDTO;
+import com.ykotsiuba.profitsoft_2.dto.author.DeleteAuthorResponseDTO;
+import com.ykotsiuba.profitsoft_2.dto.author.SaveAuthorRequestDTO;
 import com.ykotsiuba.profitsoft_2.entity.Author;
 import com.ykotsiuba.profitsoft_2.repository.AuthorRepository;
 import jakarta.transaction.Transactional;
